Distribution.Simple.SetupHooks

Description

This module defines the interface for the Hooks build-type.

To write a package that implements build-type: Hooks, you should define a module SetupHooks.hs which exports a value setupHooks :: SetupHooks. This is a record that declares actions that should be hooked into the cabal build process.

See SetupHooks for more details.

Hooks

A Cabal package with Hooks build-type must define the Haskell module SetupHooks which defines a value setupHooks :: SetupHooks.

These *setup hooks* allow package authors to customise the configuration and building of a package by providing certain hooks that get folded into the general package configuration and building logic within Cabal.

This mechanism replaces the Custom build-type, providing better integration with the rest of the Haskell ecosystem.

Usage example:

-- In your .cabal file
build-type: Hooks

custom-setup
  setup-depends:
    base        >= 4.18 && < 5,
    Cabal-hooks >= 0.1  && < 0.2

The declared Cabal version should also be at least 3.14.

-- In SetupHooks.hs, next to your .cabal file
module SetupHooks where
import Distribution.Simple.SetupHooks ( SetupHooks, noSetupHooks )

setupHooks :: SetupHooks
setupHooks =
 noSetupHooks
   { configureHooks = myConfigureHooks
   , buildHooks = myBuildHooks }

Note that SetupHooks can be monoidally combined, e.g.:

module SetupHooks where
import Distribution.Simple.SetupHooks
import qualified SomeOtherLibrary ( setupHooks )

setupHooks :: SetupHooks
setupHooks = SomeOtherLibrary.setupHooks <> mySetupHooks

mySetupHooks :: SetupHooks
mySetupHooks = ...

data SetupHooks

Hooks into the cabal build phases.

Usage:

In your .cabal file, declare build-type: Hooks (with a cabal-version greater than or equal to 3.14),

In your .cabal file, include a custom-setup stanza which declares the dependencies of your SetupHooks module; this will usually contain a dependency on the Cabal-hooks package.

Provide a SetupHooks.hs module next to your .cabal file; it must export setupHooks :: SetupHooks.

Constructors

SetupHooks

Fields

configureHooks :: ConfigureHooks

Hooks into the configure phase.

buildHooks :: BuildHooks

Hooks into the build phase.

These hooks are relevant to any build-like phase, such as repl or haddock.

installHooks :: InstallHooks

Hooks into the copy/install phase.

noSetupHooks :: SetupHooks

Empty hooks.

Configure hooks

Configure hooks can be used to augment the Cabal configure logic with package-specific logic. The main principle is that the configure hooks can feed into updating the PackageDescription of a cabal package. From then on, this package configuration is set in stone, and later hooks (e.g. hooks into the build phase) can no longer modify this configuration; instead they will receive this configuration in their inputs, and must honour it.

Configuration happens at two levels:

global configuration covers the entire package,

local configuration covers a single component.

Once the global package configuration is done, all hooks work on a per-component level. The configuration hooks thus follow a simple philosophy:

All modifications to global package options must use $sel:preConfPackageHook:ConfigureHooks.

All modifications to component configuration options must use $sel:preConfComponentHook:ConfigureHooks.

For example, to generate modules inside a given component, you should:

In the per-component configure hook, declare the modules you are going to generate by adding them to the autogenModules field for that component (unless you know them ahead of time, in which case they can be listed textually in the .cabal file of the project).

In the build hooks, describe the actions that will generate these modules.

data ConfigureHooks

Configure-time hooks.

Order of execution:

$sel:preConfPackageHook:ConfigureHooks,

configure the package,

$sel:postConfPackageHook:ConfigureHooks,

$sel:preConfComponentHook:ConfigureHooks,

configure the components.

Constructors

ConfigureHooks

Fields

preConfPackageHook :: Maybe PreConfPackageHook

Package-wide pre-configure hook. See PreConfPackageHook.

postConfPackageHook :: Maybe PostConfPackageHook

Package-wide post-configure hook. See PostConfPackageHook.

preConfComponentHook :: Maybe PreConfComponentHook

Per-component pre-configure hook. See PreConfComponentHook.

noConfigureHooks :: ConfigureHooks

Empty configure phase hooks.

Per-package configure hooks

data PreConfPackageInputs

Inputs to the package-wide pre-configure step.

Constructors

PreConfPackageInputs

Fields

configFlags :: ConfigFlags

localBuildConfig :: LocalBuildConfig

Warning: the ProgramDb in the LocalBuildInfo field will not contain any unconfigured programs.

compiler :: Compiler

platform :: Platform

data PreConfPackageOutputs

Outputs of the package-wide pre-configure step.

Prefer using noPreConfPackageOutputs and overriding the fields you care about, to avoid depending on implementation details of this datatype.

Constructors

PreConfPackageOutputs

Fields

buildOptions :: BuildOptions

extraConfiguredProgs :: ConfiguredProgs

noPreConfPackageOutputs :: PreConfPackageInputs -> PreConfPackageOutputs

Use this smart constructor to declare an empty set of changes by the package-wide pre-configure hook, and override the fields you care about.

Use this rather than PreConfPackageOutputs to avoid relying on internal implementation details of the latter.

type PreConfPackageHook = PreConfPackageInputs -> IO PreConfPackageOutputs

data PostConfPackageInputs

Inputs to the package-wide post-configure step.

Constructors

PostConfPackageInputs

Fields

localBuildConfig :: LocalBuildConfig

packageBuildDescr :: PackageBuildDescr

type PostConfPackageHook = PostConfPackageInputs -> IO ()

Package-wide post-configure step.

Perform side effects. Last opportunity for any package-wide logic; any subsequent hooks work per-component.

Per-component configure hooks

data PreConfComponentInputs

Inputs to the per-component pre-configure step.

Constructors

PreConfComponentInputs

Fields

localBuildConfig :: LocalBuildConfig

packageBuildDescr :: PackageBuildDescr

component :: Component

data PreConfComponentOutputs

Outputs of the per-component pre-configure step.

Prefer using noPreComponentOutputs and overriding the fields you care about, to avoid depending on implementation details of this datatype.

Constructors

PreConfComponentOutputs

Fields

componentDiff :: ComponentDiff

noPreConfComponentOutputs :: PreConfComponentInputs -> PreConfComponentOutputs

Use this smart constructor to declare an empty set of changes by a per-component pre-configure hook, and override the fields you care about.

Use this rather than PreConfComponentOutputs to avoid relying on internal implementation details of the latter.

type PreConfComponentHook = PreConfComponentInputs -> IO PreConfComponentOutputs

Per-component pre-configure step.

For each component of the package, this hook can perform side effects, and return a diff to the passed in component, e.g. to declare additional autogenerated modules.

newtype ComponentDiff

A diff to a Cabal Component, that gets combined monoidally into an existing Component.

Constructors

ComponentDiff

Fields

componentDiff :: Component

emptyComponentDiff :: ComponentName -> ComponentDiff

buildInfoComponentDiff :: ComponentName -> BuildInfo -> ComponentDiff

type LibraryDiff = Library

type ForeignLibDiff = ForeignLib

type ExecutableDiff = Executable

type TestSuiteDiff = TestSuite

type BenchmarkDiff = Benchmark

type BuildInfoDiff = BuildInfo

Build hooks

data BuildHooks

Build-time hooks.

Constructors

BuildHooks

Fields

preBuildComponentRules :: Maybe PreBuildComponentRules

Per-component fine-grained pre-build rules.

postBuildComponentHook :: Maybe PostBuildComponentHook

Per-component post-build hook.

noBuildHooks :: BuildHooks

Empty build hooks.

data BuildingWhat

What kind of build phase are we doing/hooking into?

Is this a normal build, or is it perhaps for running an interactive session or Haddock?

Constructors

BuildNormal BuildFlags
A normal build.

BuildRepl ReplFlags
Build steps for an interactive session.

BuildHaddock HaddockFlags
Build steps for generating documentation.

BuildHscolour HscolourFlags
Build steps for Hscolour.

buildingWhatVerbosity :: BuildingWhat -> Verbosity

buildingWhatDistPref :: BuildingWhat -> SymbolicPath Pkg ('Dir Dist)

Pre-build rules

Pre-build hooks are specified as a collection of pre-build Rules. Each Rule consists of:

a specification of its static dependencies and outputs,

the commands that execute the rule.

Rules are constructed using either one of the staticRule or dynamicRule smart constructors. Directly constructing a Rule using the constructors of that data type is not advised, as this relies on internal implementation details which are subject to change in between versions of the `Cabal-hooks` library.

Note that:

To declare the dependency on the output of a rule, one must refer to the rule directly, and not to the path to the output executing that rule will eventually produce. To do so, registering a Rule with the API returns a unique identifier for that rule, in the form of a RuleId.

File dependencies and outputs are not specified directly by FilePath, but rather use the Location type (which is more convenient when working with preprocessors).

Rules refer to the actions that execute them using static pointers, in order to enable serialisation/deserialisation of rules.

Rules can additionally monitor files or directories, which determines when to re-compute the entire set of rules.

data PreBuildComponentInputs

Constructors

PreBuildComponentInputs

Fields

buildingWhat :: BuildingWhat

what kind of build phase are we hooking into?

localBuildInfo :: LocalBuildInfo

information about the package

targetInfo :: TargetInfo

information about an individual component

type PreBuildComponentRules = Rules PreBuildComponentInputs

Post-build hooks

data PostBuildComponentInputs

Constructors

PostBuildComponentInputs

Fields

buildFlags :: BuildFlags

localBuildInfo :: LocalBuildInfo

targetInfo :: TargetInfo

type PostBuildComponentHook = PostBuildComponentInputs -> IO ()

Rules

data Rules env

A collection of Rules.

Use the rules smart constructor instead of directly using the Rules constructor.

Rules are registered using registerRule,

Monitored files or directories are declared using addRuleMonitors; a change in these will trigger the recomputation of all rules.

The env type parameter represents an extra argument, which usually consists of information known to Cabal such as LocalBuildInfo and ComponentLocalBuildInfo.

rules

Arguments

:: StaticPtr label
unique label for this collection of rules

-> (env -> RulesM ())
the computation of rules

-> Rules env

Construct a collection of rules with a given label.

A label for the rules can be constructed using the static keyword, using the StaticPointers extension. NB: separate calls to rules should have different labels.

Example usage:

myRules :: Rules env
myRules = rules (static ()) $ \ env -> do { .. } -- use the monadic API here

noRules :: RulesM ()

An empty collection of rules.

type Rule = RuleData 'User

data Dependency

A dependency of a rule.

Constructors

RuleDependency !RuleOutput
A dependency on an output of another rule.

FileDependency !Location
A direct dependency on a file at a particular location on disk.

This should not be used for files that are generated by other rules; use RuleDependency instead.

data RuleOutput

A reference to an output of another rule.

Constructors

RuleOutput

Fields

outputOfRule :: !RuleId

which rule's outputs are we referring to?

outputIndex :: !Word

which particular output of that rule?

data RuleId

A unique identifier for a Rule.

staticRule :: Typeable arg => Command arg (IO ()) -> [Dependency] -> NonEmpty Location -> Rule

A rule with static dependencies.

Prefer using this smart constructor instead of Rule whenever possible.

dynamicRule :: (Typeable depsArg, Typeable depsRes, Typeable arg) => StaticPtr (Dict (Binary depsRes, Show depsRes, Eq depsRes)) -> Command depsArg (IO ([Dependency], depsRes)) -> Command arg (depsRes -> IO ()) -> [Dependency] -> NonEmpty Location -> Rule

A rule with dynamic dependencies.

Prefer using this smart constructor instead of Rule whenever possible.

Rule inputs/outputs

Rules can declare various kinds of dependencies:

staticDependencies: files or other rules that a rule statically depends on,

extra dynamic dependencies, using the DynamicRuleCommands constructor,

MonitorFilePath: additional files and directories to monitor.

Rules are considered out-of-date precisely when any of the following conditions apply:

O1

there has been a (relevant) change in the files and directories monitored by the rules,

O2

the environment passed to the computation of rules has changed.

If the rules are out-of-date, the build system is expected to re-run the computation that computes all rules.

After this re-computation of the set of all rules, we match up new rules with old rules, by RuleId. A rule is then considered stale if any of following conditions apply:

N

the rule is new, or

S

the rule matches with an old rule, and either:

S1

a file dependency of the rule has been modifiedcreateddeleted, or a (transitive) rule dependency of the rule is itself stale, or

S2

the rule is different from the old rule, e.g. the argument stored in the rule command has changed, or the pointer to the action to run the rule has changed. (This is determined using the Eq Rule instance.)

A stale rule becomes no longer stale once we run its associated action. The build system is responsible for re-running the actions associated with each stale rule, in dependency order. This means the build system is expected to behave as follows:

Any time the rules are out-of-date, query the rules to obtain up-to-date rules.

Re-run stale rules.

data Location where

A (fully resolved) location of a dependency or result of a rule, consisting of a base directory and of a file path relative to that base directory path.

In practice, this will be something like Location dir (moduleNameSymbolicPath mod . "hs"), where:

for a file dependency, dir is one of the Cabal search directories,

for an output, dir is a directory such as autogenComponentModulesDir or componentBuildDir.

Constructors

Location

Fields

:: forall baseDir. { locationBaseDir :: !(SymbolicPath Pkg ('Dir baseDir))

Base directory.

, locationRelPath :: !(RelativePath baseDir 'File)

File path relative to base directory (including file extension).

} -> Location

location :: Location -> SymbolicPath Pkg 'File

Get a (relative or absolute) un-interpreted path to a Location.

autogenComponentModulesDir :: LocalBuildInfo -> ComponentLocalBuildInfo -> SymbolicPath Pkg ('Dir Source)

The directory in which we put auto-generated modules for a particular component.

componentBuildDir :: LocalBuildInfo -> ComponentLocalBuildInfo -> SymbolicPath Pkg ('Dir Build)

Actions

data RuleCommands (scope :: Scope) (deps :: Scope -> Type -> Type -> Type) (ruleCmd :: Scope -> Type -> Type -> Type) where

Commands to execute a rule:

for a rule with static dependencies, a single command,

for a rule with dynamic dependencies, a command for computing dynamic dependencies, and a command for executing the rule.

Constructors

StaticRuleCommand
A rule with statically-known dependencies.

Fields

:: forall arg (deps :: Scope -> Type -> Type -> Type) (ruleCmd :: Scope -> Type -> Type -> Type) (scope :: Scope). If (scope == 'System) (arg ~ ByteString) ()

=> { staticRuleCommand :: !(ruleCmd scope arg (IO ()))

The command to execute the rule.

, staticRuleArgRep :: !(If (scope == 'System) SomeTypeRep (TypeRep arg))

A TypeRep for arg.

} -> RuleCommands scope deps ruleCmd

DynamicRuleCommands

Fields

:: forall depsArg depsRes arg (deps :: Scope -> Type -> Type -> Type) (ruleCmd :: Scope -> Type -> Type -> Type) (scope :: Scope). If (scope == 'System) (depsArg ~ ByteString, depsRes ~ ByteString, arg ~ ByteString) ()

=> { dynamicRuleInstances :: !(Static scope (Dict (Binary depsRes, Show depsRes, Eq depsRes)))

A rule with dynamic dependencies, which consists of two parts:

a dynamic dependency computation, that returns additional edges to be added to the build graph together with an additional piece of data,

the command to execute the rule itself, which receives the additional piece of data returned by the dependency computation.

, dynamicDeps :: !(deps scope depsArg depsRes)

A dynamic dependency computation. The resulting dependencies will be injected into the build graph, and the result of the computation will be passed on to the command that executes the rule.

, dynamicRuleCommand :: !(ruleCmd scope arg (depsRes -> IO ()))

The command to execute the rule. It will receive the result of the dynamic dependency computation.

, dynamicRuleTypeRep :: !(If (scope == 'System) SomeTypeRep (TypeRep (depsArg, depsRes, arg)))

A TypeRep for the triple (depsArg,depsRes,arg).

} -> RuleCommands scope deps ruleCmd

type Command = CommandData 'User

A command consists of a statically-known action together with a (possibly dynamic) argument to that action.

For example, the action can consist of running an executable (such as happy or c2hs), while the argument consists of the variable component of the command, e.g. the specific file to run happy on.

mkCommand :: StaticPtr (Dict (Binary arg, Show arg)) -> StaticPtr (arg -> res) -> arg -> Command arg res

Construct a command.

Prefer using this smart constructor instead of Command whenever possible.

data Dict c where

A wrapper used to pass evidence of a constraint as an explicit value.

Constructors

Dict :: forall c. c => Dict c

Rules API

Defining pre-build rules can be done in the following style:

{-# LANGUAGE BlockArguments, StaticPointers #-}
myPreBuildRules :: PreBuildComponentRules
myPreBuildRules = rules (static ()) $ \ preBuildEnvironment -> do
  let cmd1 = mkCommand (static Dict) $ static \ arg -> do { .. }
      cmd2 = mkCommand (static Dict) $ static \ arg -> do { .. }
  myData <- liftIO someIOAction
  addRuleMonitors [ monitorDirectory "someSearchDir" ]
  registerRule_ "rule_1_1" $ staticRule (cmd1 arg1) deps1 outs1
  registerRule_ "rule_1_2" $ staticRule (cmd1 arg2) deps2 outs2
  registerRule_ "rule_1_3" $ staticRule (cmd1 arg3) deps3 outs3
  registerRule_ "rule_2_4" $ staticRule (cmd2 arg4) deps4 outs4

Here we use the rules, staticRule and mkCommand smart constructors, rather than directly using the Rules, Rule and Command constructors, which insulates us from internal changes to the Rules, Rule and Command datatypes, respectively.

We use addRuleMonitors to declare a monitored directory that the collection of rules as a whole depends on. In this case, we declare that they depend on the contents of the "searchDir" directory. This means that the rules will be computed anew whenever the contents of this directory change.

type RulesM a = RulesT IO a

Monad for constructing rules.

registerRule

Arguments

:: ShortText
user-given rule name; these should be unique on a per-package level

-> Rule
the rule to register

-> RulesM RuleId

Register a rule. Returns an identifier for that rule.

registerRule_

Arguments

:: ShortText
user-given rule name; these should be unique on a per-package level

-> Rule
the rule to register

-> RulesT IO ()

Register a rule, discarding the produced RuleId.

Using this function means that you don't expect any other rules to ever depend on any outputs of this rule. Use registerRule to retain the RuleId instead.

File/directory monitoring

addRuleMonitors :: Monad m => [MonitorFilePath] -> RulesT m ()

Declare additional monitored objects for the collection of all rules.

When these monitored objects change, the rules are re-computed.

module Distribution.Simple.FileMonitor.Types

Install hooks

data InstallHooks

Copy/install hooks.

Constructors

InstallHooks

Fields

installComponentHook :: Maybe InstallComponentHook

Per-component install hook.

noInstallHooks :: InstallHooks

Empty copy/install hooks.

data InstallComponentInputs

Constructors

InstallComponentInputs

Fields

copyFlags :: CopyFlags

localBuildInfo :: LocalBuildInfo

targetInfo :: TargetInfo

type InstallComponentHook = InstallComponentInputs -> IO ()

A per-component install hook, which can only perform side effects (e.g. copying files).

Re-exports

Hooks

Configure hooks

data ConfigFlags

Flags to configure command.

IMPORTANT: every time a new flag is added, filterConfigureFlags should be updated. IMPORTANT: every time a new flag is added, it should be added to the Eq instance

Constructors

ConfigFlags

Fields

configCommonFlags :: !CommonSetupFlags

configPrograms_ :: Option' (Last' ProgramDb)

All programs that cabal may run

configProgramPaths :: [(String, FilePath)]

user specified programs paths

configProgramArgs :: [(String, [String])]

user specified programs args

configProgramPathExtra :: NubList FilePath

Extend the $PATH

configHcFlavor :: Flag CompilerFlavor

The "flavor" of the compiler, e.g. GHC.

configHcPath :: Flag FilePath

given compiler location

configHcPkg :: Flag FilePath

given hc-pkg location

configVanillaLib :: Flag Bool

Enable vanilla library

configProfLib :: Flag Bool

Enable profiling in the library

configSharedLib :: Flag Bool

Build shared library

configStaticLib :: Flag Bool

Build static library

configDynExe :: Flag Bool

Enable dynamic linking of the executables.

configFullyStaticExe :: Flag Bool

Enable fully static linking of the executables.

configProfExe :: Flag Bool

Enable profiling in the executables.

configProf :: Flag Bool

Enable profiling in the library and executables.

configProfDetail :: Flag ProfDetailLevel

Profiling detail level in the library and executables.

configProfLibDetail :: Flag ProfDetailLevel

Profiling detail level in the library

configConfigureArgs :: [String]

Extra arguments to configure

configOptimization :: Flag OptimisationLevel

Enable optimization.

configProgPrefix :: Flag PathTemplate

Installed executable prefix.

configProgSuffix :: Flag PathTemplate

Installed executable suffix.

configInstallDirs :: InstallDirs (Flag PathTemplate)

Installation paths

configScratchDir :: Flag FilePath

configExtraLibDirs :: [SymbolicPath Pkg ('Dir Lib)]

path to search for extra libraries

configExtraLibDirsStatic :: [SymbolicPath Pkg ('Dir Lib)]

path to search for extra libraries when linking fully static executables

configExtraFrameworkDirs :: [SymbolicPath Pkg ('Dir Framework)]

path to search for extra frameworks (OS X only)

configExtraIncludeDirs :: [SymbolicPath Pkg ('Dir Include)]

path to search for header files

configIPID :: Flag String

explicit IPID to be used

configCID :: Flag ComponentId

explicit CID to be used

configDeterministic :: Flag Bool

be as deterministic as possible (e.g., invariant over GHC, database, etc). Used by the test suite

configUserInstall :: Flag Bool

The --user/--global flag

configPackageDBs :: [Maybe PackageDB]

Which package DBs to use

configGHCiLib :: Flag Bool

Enable compiling library for GHCi

configSplitSections :: Flag Bool

Enable -split-sections with GHC

configSplitObjs :: Flag Bool

Enable -split-objs with GHC

configStripExes :: Flag Bool

Enable executable stripping

configStripLibs :: Flag Bool

Enable library stripping

configConstraints :: [PackageVersionConstraint]

Additional constraints for dependencies.

configDependencies :: [GivenComponent]

The packages depended on which already exist

configPromisedDependencies :: [GivenComponent]

The packages depended on which doesn't yet exist (i.e. promised). Promising dependencies enables us to configure components in parallel, and avoids expensive builds if they are not necessary. For example, in multi-repl mode, we don't want to build dependencies that are loaded into the interactive session, since we have to build them again.

configInstantiateWith :: [(ModuleName, Module)]

The requested Backpack instantiation. If empty, either this package does not use Backpack, or we just want to typecheck the indefinite package.

configConfigurationsFlags :: FlagAssignment

configTests :: Flag Bool

Enable test suite compilation

configBenchmarks :: Flag Bool

Enable benchmark compilation

configCoverage :: Flag Bool

Enable program coverage

configLibCoverage :: Flag Bool

Enable program coverage (deprecated)

configExactConfiguration :: Flag Bool

All direct dependencies and flags are provided on the command line by the user via the '--dependency' and '--flags' options.

configFlagError :: Flag String

Halt and show an error message indicating an error in flag assignment

configRelocatable :: Flag Bool

Enable relocatable package built

configDebugInfo :: Flag DebugInfoLevel

Emit debug info.

configDumpBuildInfo :: Flag DumpBuildInfo

Should we dump available build information on build? Dump build information to disk before attempting to build, tooling can parse these files and use them to compile the source files themselves.

configUseResponseFiles :: Flag Bool

Whether to use response files at all. They're used for such tools as haddock, or ld.

configAllowDependingOnPrivateLibs :: Flag Bool

Allow depending on private sublibraries. This is used by external tools (like cabal-install) so they can add multiple-public-libraries compatibility to older ghcs by checking visibility externally.

configCoverageFor :: Flag [UnitId]

The list of libraries to be included in the hpc coverage report for testsuites run with --enable-coverage. Notably, this list must exclude indefinite libraries and instantiations because HPC does not support backpack (Nov. 2023).

Bundled Patterns

pattern ConfigCommonFlags :: Flag Verbosity -> Flag (SymbolicPath Pkg ('Dir Dist)) -> Flag (SymbolicPath CWD ('Dir Pkg)) -> Flag (SymbolicPath Pkg 'File) -> [String] -> ConfigFlags

Build hooks

data BuildFlags

Constructors

BuildFlags

Fields

buildCommonFlags :: !CommonSetupFlags

buildProgramPaths :: [(String, FilePath)]

buildProgramArgs :: [(String, [String])]

buildNumJobs :: Flag (Maybe Int)

buildUseSemaphore :: Flag String

Bundled Patterns

pattern BuildCommonFlags :: Flag Verbosity -> Flag (SymbolicPath Pkg ('Dir Dist)) -> Flag (SymbolicPath CWD ('Dir Pkg)) -> Flag (SymbolicPath Pkg 'File) -> [String] -> BuildFlags

data ReplFlags

Constructors

ReplFlags

Fields

replCommonFlags :: !CommonSetupFlags

replProgramPaths :: [(String, FilePath)]

replProgramArgs :: [(String, [String])]

replReload :: Flag Bool

replReplOptions :: ReplOptions

Bundled Patterns

pattern ReplCommonFlags :: Flag Verbosity -> Flag (SymbolicPath Pkg ('Dir Dist)) -> Flag (SymbolicPath CWD ('Dir Pkg)) -> Flag (SymbolicPath Pkg 'File) -> [String] -> ReplFlags

data HaddockFlags

Constructors

HaddockFlags

Fields

haddockCommonFlags :: !CommonSetupFlags

haddockProgramPaths :: [(String, FilePath)]

haddockProgramArgs :: [(String, [String])]

haddockHoogle :: Flag Bool

haddockHtml :: Flag Bool

haddockHtmlLocation :: Flag String

haddockForHackage :: Flag HaddockTarget

haddockExecutables :: Flag Bool

haddockTestSuites :: Flag Bool

haddockBenchmarks :: Flag Bool

haddockForeignLibs :: Flag Bool

haddockInternal :: Flag Bool

haddockCss :: Flag FilePath

haddockLinkedSource :: Flag Bool

haddockQuickJump :: Flag Bool

haddockHscolourCss :: Flag FilePath

haddockContents :: Flag PathTemplate

haddockIndex :: Flag PathTemplate

haddockKeepTempFiles :: Flag Bool

haddockBaseUrl :: Flag String

haddockLib :: Flag String

haddockOutputDir :: Flag FilePath

Bundled Patterns

pattern HaddockCommonFlags :: Flag Verbosity -> Flag (SymbolicPath Pkg ('Dir Dist)) -> Flag (SymbolicPath CWD ('Dir Pkg)) -> Flag (SymbolicPath Pkg 'File) -> [String] -> HaddockFlags

data HscolourFlags

Constructors

HscolourFlags

Fields

hscolourCommonFlags :: !CommonSetupFlags

hscolourCSS :: Flag FilePath

hscolourExecutables :: Flag Bool

hscolourTestSuites :: Flag Bool

hscolourBenchmarks :: Flag Bool

hscolourForeignLibs :: Flag Bool

Bundled Patterns

pattern HscolourCommonFlags :: Flag Verbosity -> Flag (SymbolicPath Pkg ('Dir Dist)) -> Flag (SymbolicPath CWD ('Dir Pkg)) -> Flag (SymbolicPath Pkg 'File) -> [String] -> HscolourFlags

Install hooks

data CopyFlags

Flags to copy: (destdir, copy-prefix (backwards compat), verbosity)

Constructors

CopyFlags

Fields

copyCommonFlags :: !CommonSetupFlags

copyDest :: Flag CopyDest

Bundled Patterns

pattern CopyCommonFlags :: Flag Verbosity -> Flag (SymbolicPath Pkg ('Dir Dist)) -> Flag (SymbolicPath CWD ('Dir Pkg)) -> Flag (SymbolicPath Pkg 'File) -> [String] -> CopyFlags

Hooks API

These are functions provided as part of the Hooks API. It is recommended to import them from this module as opposed to manually importing them from inside the Cabal module hierarchy.

Copy/install functions

installFileGlob

Arguments

:: Verbosity

-> CabalSpecVersion

-> Maybe (SymbolicPath CWD ('Dir Pkg))

-> (Maybe (SymbolicPath CWD ('Dir DataDir)), SymbolicPath Pkg ('Dir DataDir))
(src_dir, dest_dir)

-> RelativePath DataDir 'File
file glob pattern

-> IO ()

Install the files specified by the given glob pattern.

Interacting with the program database

data Program

Represents a program which can be configured.

Note: rather than constructing this directly, start with simpleProgram and override any extra fields.

Constructors

Program

Fields

programName :: String

The simple name of the program, eg. ghc

programFindLocation :: Verbosity -> ProgramSearchPath -> IO (Maybe (FilePath, [FilePath]))

A function to search for the program if its location was not specified by the user. Usually this will just be a call to findProgramOnSearchPath.

It is supplied with the prevailing search path which will typically just be used as-is, but can be extended or ignored as needed.

For the purpose of change monitoring, in addition to the location where the program was found, it returns all the other places that were tried.

programFindVersion :: Verbosity -> FilePath -> IO (Maybe Version)

Try to find the version of the program. For many programs this is not possible or is not necessary so it's OK to return Nothing.

programPostConf :: Verbosity -> ConfiguredProgram -> IO ConfiguredProgram

A function to do any additional configuration after we have located the program (and perhaps identified its version). For example it could add args, or environment vars.

programNormaliseArgs :: Maybe Version -> PackageDescription -> [String] -> [String]

A function that filters any arguments that don't impact the output from a commandline. Used to limit the volatility of dependency hashes when using new-build.

data ConfiguredProgram

Represents a program which has been configured and is thus ready to be run.

These are usually made by configuring a Program, but if you have to construct one directly then start with simpleConfiguredProgram and override any extra fields.

Constructors

ConfiguredProgram

Fields

programId :: String

Just the name again

programVersion :: Maybe Version

The version of this program, if it is known.

programDefaultArgs :: [String]

Default command-line args for this program. These flags will appear first on the command line, so they can be overridden by subsequent flags.

programOverrideArgs :: [String]

Override command-line args for this program. These flags will appear last on the command line, so they override all earlier flags.

programOverrideEnv :: [(String, Maybe String)]

Override environment variables for this program. These env vars will extend/override the prevailing environment of the current to form the environment for the new process.

programProperties :: Map String String

A key-value map listing various properties of the program, useful for feature detection. Populated during the configuration step, key names depend on the specific program.

programLocation :: ProgramLocation

Location of the program. eg. /usr/bin/ghc-6.4

programMonitorFiles :: [FilePath]

In addition to the programLocation where the program was found, these are additional locations that were looked at. The combination of ths found location and these not-found locations can be used to monitor to detect when the re-configuring the program might give a different result (e.g. found in a different location).

type ProgArg = String

data ProgramLocation

Where a program was found. Also tells us whether it's specified by user or not. This includes not just the path, but the program as well.

Constructors

UserSpecified
The user gave the path to this program, eg. --ghc-path=/usr/bin/ghc-6.6

Fields

locationPath :: FilePath

FoundOnSystem
The program was found automatically.

Fields

locationPath :: FilePath

data ProgramDb

The configuration is a collection of information about programs. It contains information both about configured programs and also about programs that we are yet to configure.

The idea is that we start from a collection of unconfigured programs and one by one we try to configure them at which point we move them into the configured collection. For unconfigured programs we record not just the Program but also any user-provided arguments and location for the program.

addKnownPrograms :: [Program] -> ProgramDb -> ProgramDb

configureUnconfiguredProgram :: Verbosity -> Program -> ProgramDb -> IO (Maybe ConfiguredProgram)

Try to configure a specific program. If the program is already included in the collection of unconfigured programs then we use any user-supplied location and arguments.

simpleProgram :: String -> Program

Make a simple named program.

By default we'll just search for it in the path and not try to find the version name. You can override these behaviours if necessary, eg:

(simpleProgram "foo") { programFindLocation = ... , programFindVersion ... }

General Cabal datatypes

data Verbosity

data Compiler

Constructors

Compiler

Fields

compilerId :: CompilerId

Compiler flavour and version.

compilerAbiTag :: AbiTag

Tag for distinguishing incompatible ABI's on the same architecture/os.

compilerCompat :: [CompilerId]

Other implementations that this compiler claims to be compatible with.

compilerLanguages :: [(Language, CompilerFlag)]

Supported language standards.

compilerExtensions :: [(Extension, Maybe CompilerFlag)]

Supported extensions.

compilerProperties :: Map String String

A key-value map for properties not covered by the above fields.

data Platform

Constructors

Platform Arch OS

newtype Suffix

A suffix (or file extension).

Mostly used to decide which preprocessor to use, e.g. files with suffix "y" are usually processed by the "happy" build tool.

Constructors

Suffix String

Package information

data LocalBuildConfig

LocalBuildConfig contains options that can be controlled by the user and serve as inputs to the configuration of a package.

data LocalBuildInfo

Data cached after configuration step. See also ConfigFlags.

data PackageBuildDescr

PackageBuildDescr contains the information Cabal determines after performing package-wide configuration of a package, before doing any per-component configuration.

data PackageDescription

This data type is the internal representation of the file pkg.cabal. It contains two kinds of information about the package: information which is needed for all packages, such as the package name and version, and information which is needed for the simple build system only, such as the compiler options and library name.

Constructors

PackageDescription

Fields

specVersion :: CabalSpecVersion

The version of the Cabal spec that this package description uses.

package :: PackageIdentifier

licenseRaw :: Either License License

licenseFiles :: [RelativePath Pkg 'File]

copyright :: !ShortText

maintainer :: !ShortText

author :: !ShortText

stability :: !ShortText

testedWith :: [(CompilerFlavor, VersionRange)]

homepage :: !ShortText

pkgUrl :: !ShortText

bugReports :: !ShortText

sourceRepos :: [SourceRepo]

synopsis :: !ShortText

A one-line summary of this package

description :: !ShortText

A more verbose description of this package

category :: !ShortText

customFieldsPD :: [(String, String)]

Custom fields starting with x-, stored in a simple assoc-list.

buildTypeRaw :: Maybe BuildType

The original build-type value as parsed from the .cabal file without defaulting. See also buildType.

Since: Cabal-syntax-2.2

setupBuildInfo :: Maybe SetupBuildInfo

library :: Maybe Library

subLibraries :: [Library]

executables :: [Executable]

foreignLibs :: [ForeignLib]

testSuites :: [TestSuite]

benchmarks :: [Benchmark]

dataFiles :: [RelativePath DataDir 'File]

data file globs, relative to data directory

dataDir :: SymbolicPath Pkg ('Dir DataDir)

data directory (may be absolute, or relative to package)

extraSrcFiles :: [RelativePath Pkg 'File]

extraTmpFiles :: [RelativePath Pkg 'File]

extraDocFiles :: [RelativePath Pkg 'File]

Component information

data Component

Constructors

CLib Library

CFLib ForeignLib

CExe Executable

CTest TestSuite

CBench Benchmark

data ComponentName

Constructors

CLibName LibraryName

CNotLibName NotLibComponentName

Bundled Patterns

pattern CBenchName :: UnqualComponentName -> ComponentName

pattern CTestName :: UnqualComponentName -> ComponentName

pattern CExeName :: UnqualComponentName -> ComponentName

pattern CFLibName :: UnqualComponentName -> ComponentName

componentName :: Component -> ComponentName

data BuildInfo

Constructors

BuildInfo

Fields

buildable :: Bool

component is buildable here

buildTools :: [LegacyExeDependency]

Tools needed to build this bit.

This is a legacy field that buildToolDepends largely supersedes.

Unless use are very sure what you are doing, use the functions in Distribution.Simple.BuildToolDepends rather than accessing this field directly.

buildToolDepends :: [ExeDependency]

Haskell tools needed to build this bit

This field is better than buildTools because it allows one to precisely specify an executable in a package.

Unless use are very sure what you are doing, use the functions in Distribution.Simple.BuildToolDepends rather than accessing this field directly.

cppOptions :: [String]

options for pre-processing Haskell code

asmOptions :: [String]

options for assembler

cmmOptions :: [String]

options for C-- compiler

ccOptions :: [String]

options for C compiler

cxxOptions :: [String]

options for C++ compiler

ldOptions :: [String]

options for linker

hsc2hsOptions :: [String]

options for hsc2hs

pkgconfigDepends :: [PkgconfigDependency]

pkg-config packages that are used

frameworks :: [RelativePath Framework 'File]

support frameworks for Mac OS X

extraFrameworkDirs :: [SymbolicPath Pkg ('Dir Framework)]

extra locations to find frameworks.

asmSources :: [SymbolicPath Pkg 'File]

Assembly files.

cmmSources :: [SymbolicPath Pkg 'File]

C-- files.

cSources :: [SymbolicPath Pkg 'File]

cxxSources :: [SymbolicPath Pkg 'File]

jsSources :: [SymbolicPath Pkg 'File]

hsSourceDirs :: [SymbolicPath Pkg ('Dir Source)]

where to look for the Haskell module hierarchy

otherModules :: [ModuleName]

non-exposed or non-main modules

virtualModules :: [ModuleName]

exposed modules that do not have a source file (e.g. GHC.Prim from ghc-prim package)

autogenModules :: [ModuleName]

not present on sdist, Paths_* or user-generated with a custom Setup.hs

defaultLanguage :: Maybe Language

language used when not explicitly specified

otherLanguages :: [Language]

other languages used within the package

defaultExtensions :: [Extension]

language extensions used by all modules

otherExtensions :: [Extension]

other language extensions used within the package

oldExtensions :: [Extension]

the old extensions field, treated same as defaultExtensions

extraLibs :: [String]

what libraries to link with when compiling a program that uses your package

extraLibsStatic :: [String]

what libraries to link with when compiling a program fully statically that uses your package

extraGHCiLibs :: [String]

if present, overrides extraLibs when package is loaded with GHCi.

extraBundledLibs :: [String]

if present, adds libs to hs-libraries, which become part of the package. Example 1: the Cffi library shipping with the rts, alongside the HSrts-1.0.a,.o,... Example 2: a library that is being built by a foreign tool (e.g. rust) and copied and registered together with this library. The logic on how this library is built will have to be encoded in a custom Setup for now. Otherwise cabal would need to learn how to call arbitrary library builders.

extraLibFlavours :: [String]

Hidden Flag. This set of strings, will be appended to all libraries when copying. E.g. [libHSname_flavour | flavour <- extraLibFlavours]. This should only be needed in very specific cases, e.g. the rts package, where there are multiple copies of slightly differently built libs.

extraDynLibFlavours :: [String]

Hidden Flag. This set of strings will be appended to all dynamic libraries when copying. This is particularly useful with the rts package, where we want different dynamic flavours of the RTS library to be installed.

extraLibDirs :: [SymbolicPath Pkg ('Dir Lib)]

extraLibDirsStatic :: [SymbolicPath Pkg ('Dir Lib)]

includeDirs :: [SymbolicPath Pkg ('Dir Include)]

directories to find .h files

includes :: [SymbolicPath Include 'File]

The .h files to be found in includeDirs

autogenIncludes :: [RelativePath Include 'File]

The .h files to be generated (e.g. by autoconf)

installIncludes :: [RelativePath Include 'File]

.h files to install with the package

options :: PerCompilerFlavor [String]

profOptions :: PerCompilerFlavor [String]

sharedOptions :: PerCompilerFlavor [String]

staticOptions :: PerCompilerFlavor [String]

customFieldsBI :: [(String, String)]

Custom fields starting with x-, stored in a simple assoc-list.

targetBuildDepends :: [Dependency]

Dependencies specific to a library or executable target

mixins :: [Mixin]

emptyBuildInfo :: BuildInfo

data TargetInfo

The TargetInfo contains all the information necessary to build a specific target (e.g., componentmodulefile) in a package. In principle, one can get the Component from a ComponentLocalBuildInfo and LocalBuildInfo, but it is much more convenient to have the component in hand.

Constructors

TargetInfo

Fields

targetCLBI :: ComponentLocalBuildInfo

targetComponent :: Component

data ComponentLocalBuildInfo

The first five fields are common across all algebraic variants.

Constructors

LibComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentIsIndefinite_ :: Bool

Is this an indefinite component (i.e. has unfilled holes)?

componentInstantiatedWith :: [(ModuleName, OpenModule)]

How the component was instantiated

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

componentCompatPackageKey :: String

Compatibility "package key" that we pass to older versions of GHC.

componentCompatPackageName :: MungedPackageName

Compatibility "package name" that we register this component as.

componentExposedModules :: [ExposedModule]

A list of exposed modules (either defined in this component, or reexported from another component.)

componentIsPublic :: Bool

Convenience field, specifying whether or not this is the "public library" that has the same name as the package.

FLibComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

ExeComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

TestComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

BenchComponentLocalBuildInfo

Fields

componentLocalName :: ComponentName

It would be very convenient to store the literal Library here, but if we do that, it will get serialized (via the Binary) instance twice. So instead we just provide the ComponentName, which can be used to find the Component in the PackageDescription. NB: eventually, this will NOT uniquely identify the ComponentLocalBuildInfo.

componentComponentId :: ComponentId

The computed ComponentId of this component.

componentUnitId :: UnitId

The computed UnitId which uniquely identifies this component. Might be hashed.

componentPackageDeps :: [(UnitId, MungedPackageId)]

Resolved internal and external package dependencies for this component. The BuildInfo specifies a set of build dependencies that must be satisfied in terms of version ranges. This field fixes those dependencies to the specific versions available on this machine for this compiler.

componentIncludes :: [(OpenUnitId, ModuleRenaming)]

The set of packages that are brought into scope during compilation, including a ModuleRenaming which may used to hide or rename modules. This is what gets translated into -package-id arguments. This is a modernized version of componentPackageDeps, which is kept around for BC purposes.

componentExeDeps :: [UnitId]

componentInternalDeps :: [UnitId]

The internal dependencies which induce a graph on the ComponentLocalBuildInfo of this package. This does NOT coincide with componentPackageDeps because it ALSO records 'build-tool' dependencies on executables. Maybe one day cabal-install will also handle these correctly too!

Components

data Library

Constructors

Library

Fields

libName :: LibraryName

exposedModules :: [ModuleName]

reexportedModules :: [ModuleReexport]

signatures :: [ModuleName]

What sigs need implementations?

libExposed :: Bool

Is the lib to be exposed by default? (i.e. whether its modules available in GHCi for example)

libVisibility :: LibraryVisibility

Whether this multilib can be used as a dependency for other packages.

libBuildInfo :: BuildInfo

data ForeignLib

A foreign library stanza is like a library stanza, except that the built code is intended for consumption by a non-Haskell client.

Constructors

ForeignLib

Fields

foreignLibName :: UnqualComponentName

Name of the foreign library

foreignLibType :: ForeignLibType

What kind of foreign library is this (static or dynamic).

foreignLibOptions :: [ForeignLibOption]

What options apply to this foreign library (e.g., are we merging in all foreign dependencies.)

foreignLibBuildInfo :: BuildInfo

Build information for this foreign library.

foreignLibVersionInfo :: Maybe LibVersionInfo

Libtool-style version-info data to compute library version. Refer to the libtool documentation on the current:revision:age versioning scheme.

foreignLibVersionLinux :: Maybe Version

Linux library version

foreignLibModDefFile :: [RelativePath Source 'File]

(Windows-specific) module definition files

This is a list rather than a maybe field so that we can flatten the condition trees (for instance, when creating an sdist)

data Executable

Constructors

Executable

Fields

exeName :: UnqualComponentName

modulePath :: RelativePath Source 'File

exeScope :: ExecutableScope

buildInfo :: BuildInfo

data TestSuite

A "test-suite" stanza in a cabal file.

Constructors

TestSuite

Fields

testName :: UnqualComponentName

testInterface :: TestSuiteInterface

testBuildInfo :: BuildInfo

testCodeGenerators :: [String]

data Benchmark

A "benchmark" stanza in a cabal file.

Constructors

Benchmark

Fields

benchmarkName :: UnqualComponentName

benchmarkInterface :: BenchmarkInterface

benchmarkBuildInfo :: BuildInfo

data LibraryName

Constructors

LMainLibName

LSubLibName UnqualComponentName

emptyLibrary :: Library

emptyForeignLib :: ForeignLib

An empty foreign library.

emptyExecutable :: Executable

emptyTestSuite :: TestSuite

emptyBenchmark :: Benchmark